+++ indico/modules/events/papers/forms.py
@@ -20,7 +20,7 @@ from indico.web.forms.fields import (EditableFileField, FileField, HiddenEnumFie
                                      IndicoDateTimeField, IndicoMarkdownField, IndicoTagListField)
 from indico.web.forms.fields.principals import PrincipalListField
 from indico.web.forms.util import inject_validators
-from indico.web.forms.validators import HiddenUnless, LinkedDateTime, UsedIf
+from indico.web.forms.validators import HiddenUnless, LinkedDateTime
 from indico.web.forms.widgets import SwitchWidget
 
 
@@ -83,11 +83,9 @@ class BulkPaperJudgmentForm(IndicoForm):
 class PaperReviewingSettingsForm(IndicoForm):
     """Settings form for paper reviewing."""
 
-    RATING_FIELDS = ('scale_lower', 'scale_upper')
-
     announcement = IndicoMarkdownField(_('Announcement'), editor=True)
-    scale_lower = IntegerField(_("Scale (from)"), [UsedIf(lambda form, field: not form.has_ratings), InputRequired()])
-    scale_upper = IntegerField(_("Scale (to)"), [UsedIf(lambda form, field: not form.has_ratings), InputRequired()])
+    scale_lower = IntegerField(_("Scale (from)"), [InputRequired()])
+    scale_upper = IntegerField(_("Scale (to)"), [InputRequired()])
     email_settings = PaperEmailSettingsField(_("Email notifications"))
 
     def __init__(self, *args, **kwargs):
@@ -95,8 +93,7 @@ class PaperReviewingSettingsForm(IndicoForm):
         self.has_ratings = kwargs.pop('has_ratings', False)
         super().__init__(*args, **kwargs)
         if self.has_ratings:
-            self.scale_upper.warning = _("Some reviewers have already submitted ratings so the scale cannot be changed "
-                                         "anymore.")
+            self.scale_upper.warning = _("Changing the ratings scale will proportionally affect existing ratings.")
 
     def validate_scale_upper(self, field):
         lower = self.scale_lower.data
@@ -107,14 +104,6 @@ class PaperReviewingSettingsForm(IndicoForm):
             raise ValidationError(_("The scale's 'to' value must be greater than the 'from' value."))
         if upper - lower > 20:
             raise ValidationError(_("The difference between 'to' and' from' may not be greater than 20."))
-
-    @property
-    def data(self):
-        data = super().data
-        if self.has_ratings:
-            for key in self.RATING_FIELDS:
-                del data[key]
-        return data
 
 
 class PaperSubmissionForm(IndicoForm):